+++ react/Conditions.js
@@ -1,4 +1,0 @@
-/* eslint-disable camelcase */
-import EXPERIMENTAL_Conditions from './components/EXPERIMENTAL_Conditions/index'
-
-export default EXPERIMENTAL_Conditions