+++ react/Conditions.ts
@@ -1,0 +1,3 @@
+import Conditions from './components/Conditions/index'
+
+export default Conditions