+++ react/components/Conditions/typings.d.ts
@@ -1,0 +1,12 @@
+export type Labels = {
+  addNewCondition?: string
+  delete?: string
+  operatorAll?: string
+  operatorAnd?: string
+  operatorAny?: string
+  operatorOr?: string
+  headerPrefix?: string
+  headerSufix?: string
+}
+
+export type Operator = 'all' | 'any'